Include all observation from the sciplex data
The first notebook simply does the gene matching with the lincs data, ignoring the subsetting from before. The second notebook is anupdated version of the addition of SMILES strings to the `.obs`  dataframe. As a result all `.h5ad` files in the `PROJECT_DIR/'datasets'` folder are updated and ready to be used in our model sweeps.