Include all observations from the sciplex data (#65)
* Include all observation from the sciplex data

The first notebook simply does the gene matching with the lincs data, ignoring the subsetting from before. The second notebook is anupdated version of the addition of SMILES strings to the `.obs`  dataframe. As a result all `.h5ad` files in the `PROJECT_DIR/'datasets'` folder are updated and ready to be used in our model sweeps.

* Fix JQ1

* Remove redundant code

The cell which renamed the "JQ1" a second time was removed.

Co-authored-by: Simon Boehm <simon@siboehm.com>